reduce interfaces
