+++ components/common/SpaceAccessGate/components/TokenGate/hooks/useTokenGates.ts
@@ -8,7 +8,8 @@ import { useUser } from 'hooks/useUser';
 import { useWeb3Account } from 'hooks/useWeb3Account';
 import type { AuthSig } from 'lib/blockchain/interfaces';
 import type { SpaceWithGates } from 'lib/spaces/interfaces';
-import type { TokenGateEvaluationResult, TokenGateJoinType, TokenGateWithRoles } from 'lib/tokenGates/interfaces';
+import type { TokenGateEvaluationResult } from 'lib/tokenGates/evaluateEligibility';
+import type { TokenGateJoinType, TokenGateWithRoles } from 'lib/tokenGates/interfaces';
 import { lowerCaseEqual } from 'lib/utilities/strings';
 
 type Props = {
