+++ lib/tokenGates/interfaces.ts
@@ -1,28 +1,9 @@
-import type { Role, Space, TokenGate, TokenGateToRole } from '@charmverse/core/prisma';
+import type { Role, TokenGate, TokenGateToRole } from '@charmverse/core/prisma';
 
 export type TokenGateJoinType = 'public_bounty_token_gate' | 'token_gate';
 
 export interface TokenGateWithRoles extends TokenGate {
   tokenGateToRoles: (TokenGateToRole & { role: Role })[];
-}
-
-export type TokenGateJwtResult = { jwt?: string; id: string; verified: boolean; grantedRoles: string[] };
-
-export interface TokenGateJwt {
-  signedToken: string;
-  tokenGate: TokenGateWithRoles;
-}
-
-/**
- * @gateTokens List of Lit-generated tokens we can verify when joining a space
- */
-export interface TokenGateEvaluationResult {
-  userId: string;
-  space: Space;
-  walletAddress: string;
-  canJoinSpace: boolean;
-  gateTokens: TokenGateJwt[];
-  roles: Role[];
 }
 
 export type TokenGateAccessType =
